Add Arabic language and make Settings screen scrollable.
- Add Arabic language, the app surprisingly works very well with Right-To-Left (RTL) languages 👍
- Make `Settings` screen scrollable using swipe gesture or mouse wheel.

Before this, if the screen is too small or the user applied text enlargement
from Android's Accessibility. Then some items in the `Settings` screen, well,
might not be shown and be off the screen.

Now it doesn't care about the screen size, the user can scroll using
swipe gesture / mouse.scroll wheel.