Fix notifications.
If the opened file URIs list is empty then check again after grace period.
If list is empty cancel work. otherwise, most notification gets canceled
as soon as opened

This is to ensure there is no sync issues between threads,
the list is reported empty when it is clearly not empty! But when we
check again after few seconds, the list correctly reports the opened
files. If it was indeed empty, then cancel work by
`deferredUntilCompleted.complete(Unit)`.

I opted to use `Handler(Looper.getMainLooper()).postDelayed()` instead of `Timer`.
The Handler uses the main applications thread, while the Timer creates a background
thread to call the deferred work. I didn't notice much difference between both,
but Handler is safer if we are checking the value of StateFlow
`cifsRepository.openUriList.value` since if we did that in a background thread,
that might cause race conditions and sync issues between threads.

And since WorkManager automatically runs it on a background thread, that might
have caused a race condition when checking the state flow. This is just a guess.

Anyways. This fixes the notifications, and has no bad side effects after extensive testing.